+++ java/src/test/java/io/cucumber/prettyformatter/MessagesToSummaryWriterTest.java
@@ -15,30 +15,36 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-class MessagesToPrettyWriterTest {
+class MessagesToSummaryWriterTest {
 
     @Test
-    void it_writes_two_messages_to_pretty() throws IOException {
+    void it_writes_two_messages_to_summary() throws IOException {
         Instant started = Instant.ofEpochSecond(10);
         Instant finished = Instant.ofEpochSecond(30);
 
-        String out = renderAsPretty(
+        String out = renderAsSummary(
                 Envelope.of(new TestRunStarted(toMessage(started), "some-id")),
                 Envelope.of(new TestRunFinished(null, true, toMessage(finished), null, "some-id")));
 
-        assertThat(out).isEmpty();
+        assertThat(out).isEqualToNormalizingNewlines("\n" +
+                "0 scenarios\n" +
+                "0 steps\n" +
+                "0m 20.0s\n"
+        );
     }
 
     @Test
-    void it_writes_no_message_to_pretty() throws IOException {
-        String out = renderAsPretty();
-        assertThat(out).isEmpty();
+    void it_writes_no_message_to_summary() throws IOException {
+        String out = renderAsSummary();
+        assertThat(out).isEqualToNormalizingNewlines("\n" +
+                "0 scenarios\n" +
+                "0 steps\n");
     }
 
     @Test
     void it_throws_when_writing_after_close() {
         ByteArrayOutputStream bytes = new ByteArrayOutputStream();
-        MessagesToPrettyWriter writer = create(bytes);
+        MessagesToSummaryWriter writer = create(bytes);
         writer.close();
         assertThrows(IOException.class, () -> writer.write(null));
     }
@@ -46,14 +52,14 @@ class MessagesToPrettyWriterTest {
     @Test
     void it_can_be_closed_twice() {
         ByteArrayOutputStream bytes = new ByteArrayOutputStream();
-        MessagesToPrettyWriter writer = create(bytes);
+        MessagesToSummaryWriter writer = create(bytes);
         writer.close();
         assertDoesNotThrow(writer::close);
     }
 
-    private static String renderAsPretty(Envelope... messages) throws IOException {
+    private static String renderAsSummary(Envelope... messages) throws IOException {
         ByteArrayOutputStream bytes = new ByteArrayOutputStream();
-        try (MessagesToPrettyWriter writer = create(bytes)) {
+        try (MessagesToSummaryWriter writer = create(bytes)) {
             for (Envelope message : messages) {
                 writer.write(message);
             }
@@ -62,7 +68,7 @@ class MessagesToPrettyWriterTest {
         return new String(bytes.toByteArray(), UTF_8);
     }
 
-    private static MessagesToPrettyWriter create(ByteArrayOutputStream bytes) {
-        return MessagesToPrettyWriter.builder().build(bytes);
+    private static MessagesToSummaryWriter create(ByteArrayOutputStream bytes) {
+        return MessagesToSummaryWriter.builder().build(bytes);
     }
 }